Split window aggregate tests
